+++ src/utils/common.js
@@ -1,0 +1,9 @@
+const getRandomArrayElement = (items) =>
+  items[Math.floor(Math.random() * items.length)];
+
+const getRandomNumber = () => Math.floor(Math.random() * 100);
+
+export {
+  getRandomArrayElement,
+  getRandomNumber,
+};
